requirements.txt
fastapi==0.128.6
uvicorn[standard]==0.40.0

sqlalchemy==2.0.48
asyncpg==0.31.0

pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
email-validator==2.3.0

pyjwt==2.12.0
pwdlib[argon2]==0.3.0

loguru==0.7.3